Enemy Sprite Change & Added Attack State
I was too lazy to animate an attack for the zombie sprite I was using, so I found a new asset that includes an attack animation - a zombie. Putting in the zombie messed up the movement for a while because I didn't set the origin for the collision mask, but now it works normally again.

Then attack an attack state so the zombie attacks within a certain range.

Gotta push, that was a headache!