Added List Data Structure to track hitboxes
Used a list data structure to keep track of objects the hitbox has already touched - this was to solve the problem where the player was getting hit by the enemy four times per attack. 

After adding this, I lost the last frame of the attack animation for the player. Was able to solve (thanks dad) by adding another duplicate last frame to the player attack animation.