Added Knockback State
Took me around an hour to get the knockback implemented correctly. Added an approach script as a linear alternative to lerp(). 

The knockback is working right now for the skeleton (enemy), but not for the player.

Pushing this crud because it took me so long to get the skeleton working correctly. I broke it a few times. Eh, typical!