More HUD Elements, Player Variables Updated
When adding kills and level to the HUD when building my leveling system, I noticed that these variables were being reset in every room. I realized that was because I was creating a new instance of the player in every room rather than teleporting the first instance to the next room. In order to avoid creating teleportation triggers for moving the same instance to the next room (I have no idea how to do that soooo -> improvise!), I created new global characters that would set as their default values when the current room == 1, meaning that they were in the first room after the menu screen. I think a pretty solid fix to have implemented in < 10 minutes aha!