Procedural Generation Scripts
Created scripts to handle placing platforms of different sizes within my other procedural generation scripts/controllers. I realized that it was problematic when choosing a random platform to place, that I would also have to increment the x-coordinate of the controller by the length of the platform. To make this simple, I just created scripts for each platform length that would both place the platform then increment the controller the length of the platform.