Fixed Random_platforms()
For some reason, when using choose() and listing all the platforms within, it would only choose one of the first two of the functions placed inside. Not sure if that is a gamemaker bug or if it doesn't want to choose a function... But to fix it, I just use a switch statement instead and generated a random int within the range of platforms I have in order to select a case. This one works much better AKA actually works correctly.